# Speeds up tests