--- a/pyproject.toml
+++ b/pyproject.toml
@@ -52,6 +52,7 @@ setuptools = { version = ">=17.0.0", python = "<3.8" }
 pytest-cov = "^6.1.1"
 pytest-datadir = "^1.5.0"
 pytest-mock = "^3.14.0"
+pytest-xdist = "^3.6.1"
 pycodestyle = ">=2.9.1"
 flake8 = "^5.0.4"
 pyflakes = "^2.5.0"