Speeds up tests. Ain't Nobody Got Time For That...

Old: `pytest --cov --cov-report=xml  6.01s user 1.46s system 86% cpu 8.588 total`
New: `pytest --cov --cov-report=xml -n auto  17.06s user 3.83s system 428% cpu 4.875 total`

or 1.76x faster!