@Bachmann1234 2025-05-11
Very nice! Thanks for the pr. This I can merge now